Fixed healthchecks not working and cleaned up exampleRequests
